requirements.txt
# Google Gemini API (free tier)
google-generativeai>=0.8.0

# Data models
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# For local testing (optional)
jupyter>=1.0.0
pytest>=7.0.0